fix(DatasetDiff): fixes false "no changes detected" returns
Previously, if you only had Viz or Transform in one dataset, not both, DatasetDiff would not attempt to diff Viz or Transform. This commit fixes that error, allowing you to add a viz or transform without having to make any other changes to a dataset.

This also updates the test cases on DatasetDiff, and makes sure they follow the latest dataset formatting (DataPath => BodyPath, VisConfig => Viz)